refactor: use shared HorizontalScroll component for Chips and Calendar
