+++ src/common/Chips/Chips.tsx
@@ -1,7 +1,7 @@
 // Copyright (C) 2017-2024 Smart code 203358507
 
-import React, { memo, useEffect, useRef, useState } from 'react';
-import classNames from 'classnames';
+import React, { memo } from 'react';
+import HorizontalScroll from '../HorizontalScroll';
 import Chip from './Chip';
 import styles from './Chips.less';
 
@@ -16,28 +16,9 @@ type Props = {
     onSelect: (value: string) => {},
 };
 
-const SCROLL_THRESHOLD = 1;
-
 const Chips = memo(({ options, selected, onSelect }: Props) => {
-    const ref = useRef<HTMLDivElement>(null);
-    const [scrollPosition, setScrollPosition] = useState('left');
-
-    useEffect(() => {
-        const onScroll = ({ target }: Event) => {
-            const { scrollLeft, scrollWidth, offsetWidth} = target as HTMLDivElement;
-            const position =
-                (scrollLeft - SCROLL_THRESHOLD) <= 0 ? 'left' :
-                    (scrollLeft + offsetWidth + SCROLL_THRESHOLD) >= scrollWidth ? 'right' :
-                        'center';
-            setScrollPosition(position);
-        };
-
-        ref.current?.addEventListener('scroll', onScroll);
-        return () => ref.current?.removeEventListener('scroll', onScroll);
-    }, []);
-
     return (
-        <div ref={ref} className={classNames(styles['chips'], [styles[scrollPosition]])}>
+        <HorizontalScroll className={styles['chips']}>
             {
                 options.map(({ label, value }) => (
                     <Chip
@@ -49,7 +30,7 @@ const Chips = memo(({ options, selected, onSelect }: Props) => {
                     />
                 ))
             }
-        </div>
+        </HorizontalScroll>
     );
 });
