+++ src/routes/Calendar/Table/Cell/Cell.less
@@ -45,7 +45,6 @@
         gap: 1rem;
         height: 0;
         padding: 1rem;
-        overflow-x: scroll;
 
         .item {
             flex: none;
